src/main/java/micha/udemy/recipeapp/converter/UnitOfMeasureCommandToUnitOfMeasure.java
package micha.udemy.recipeapp.converter;

import lombok.Synchronized;
import micha.udemy.recipeapp.command.UnitOfMeasureCommand;
import micha.udemy.recipeapp.model.UnitOfMeasure;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Controller;

@Controller
public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand, UnitOfMeasure> {

    @Synchronized
    @Nullable
    @Override
    public UnitOfMeasure convert(UnitOfMeasureCommand source) {
        if (source == null) {
            return null;
        }

        final UnitOfMeasure uom = new UnitOfMeasure();
        uom.setId(source.getId());
        uom.setDescription(source.getDescription());
        return uom;
    }

}
